fix(tooling): bound source governance probes
