+++ scripts/code-complexity-budget.test.mjs
@@ -18,6 +18,7 @@ import {
   classify,
   composeRenameEvidence,
   dispositionSoftWarnings,
+  git,
   hasGeneratedProvenance,
   isEligible,
   ownerFor,
@@ -46,6 +47,17 @@ const groups = {
   },
   'generated-projection:javascript-typescript': { hard: 100 },
 };
+
+test('git probes fail fast with a precise timeout', () => {
+  assert.throws(
+    () =>
+      git(['show', 'HEAD:file'], {}, () => ({
+        status: null,
+        error: { code: 'ETIMEDOUT' },
+      })),
+    /timed out after 10000ms/,
+  );
+});
 
 test('classification order covers every declared source class', () => {
   assert.equal(
